Online messaging application for the visually impaired
This project aims to develop a user-friendly and accessible messaging application for
visually impaired and non-disabled individuals. The main purpose of the project is to
communicate with voice commands without the need for a visual interface. The application
can be used with both voice commands and traditional keyboard text messaging methods,
providing the user with a versatile communication tool.
The uniqueness of this project lies in the fact that it enables the effective use of voice
commands and the reading of incoming messages out loud for visually impaired users. Thanks
to these features, visually impaired individuals can communicate without difficulty in the
messaging process. Additionally, the application offers a modern and accessible interface for
all users.